Refactors parse_spec script into individual classes
